fix: adjusting the Bookmarks Directory
