+++ src/main/java/project/linkarchive/backend/bookmark/controller/BookMarkQueryController.java
@@ -1,5 +1,7 @@
 package project.linkarchive.backend.bookmark.controller;
 
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import project.linkarchive.backend.bookmark.service.BookMarkQueryService;
 import project.linkarchive.backend.hashtag.response.TagListResponse;
+import project.linkarchive.backend.link.response.linkList.UserLinkListResponse;
 
 @RestController
 @PreAuthorize("isAuthenticated()")
@@ -17,6 +20,17 @@ public class BookMarkQueryController {
 
     public BookMarkQueryController(BookMarkQueryService bookMarkQueryService) {
         this.bookMarkQueryService = bookMarkQueryService;
+    }
+
+    // 사용자별 북마크 리스트 조회 010
+    @GetMapping("/mark/links/user/{userId}")
+    public ResponseEntity<UserLinkListResponse> getMarkedLinkList(
+            @PathVariable("userId") Long userId,
+            @RequestParam(value = "linkId", required = false) Long lastLinkId,
+            @PageableDefault Pageable pageable
+    ) {
+        UserLinkListResponse userLinkListResponse = bookMarkQueryService.getMarkedLinkList(userId, lastLinkId, pageable);
+        return ResponseEntity.ok(userLinkListResponse);
     }
 
     // 사용자 별 북마크 리스트 중 해시태그 리스트 조회 011
